backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────
fastapi
uvicorn[standard]
python-multipart

# ── Database ──────────────────────────────────────────────────────
psycopg2-binary
sqlalchemy

# ── LLM & LangChain ──────────────────────────────────────────────
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-chroma
langchain-openai              # OpenAI LLM + embeddings
langgraph

# ── HuggingFace inference (API only, no local models) ────────────
huggingface-hub

# ── Vector store ──────────────────────────────────────────────────
chromadb

# ── PDF processing ────────────────────────────────────────────────
pypdf

# ── Utilities ─────────────────────────────────────────────────────
python-dotenv
pydantic
